+++ internal/service/model.go
@@ -1,7 +1,7 @@
 package todo_service
 
 type ToDo struct {
-	Id uint64
+	Id int32
 	Name string
 	Text string 
 }
@@ -9,7 +9,7 @@ type ToDo struct {
 
 type ToDos []*ToDo
 
-func (t ToDos) FilterById(id uint64) *ToDo {
+func (t ToDos) FilterById(id int32) *ToDo {
 	for _, to := range t{
 		if to.Id == id{
 			return to
